+++ 2_DN/sic_xe_simulator/src/machine/devices/output_device.rs
@@ -6,9 +6,9 @@ pub struct OutputDevice {}
 impl Device for OutputDevice {
     fn test(&self) -> bool { true }
 
-    fn read(&mut self) -> i8 { 0 }
+    fn read(&mut self) -> u8 { 0 }
 
-    fn write(&mut self, val: i8) -> () {
-        let _ = io::stdout().write_all(&[val as u8]).expect("Stdout error");
+    fn write(&mut self, val: u8) -> () {
+        let _ = io::stdout().write_all(&[val]).expect("Stdout error");
     }
 }
